+++ eng/tools/generator/cmd/automation/pipeline/model.go
@@ -18,6 +18,7 @@ type GenerateInput struct {
 	RepoHTTPSURL            string                        `json:"repoHttpsUrl,omitempty"`
 	Trigger                 string                        `json:"trigger,omitempty"`
 	ChangedFiles            []string                      `json:"changedFiles,omitempty"`
+	RelatedReadmeMdFile     string                        `json:"relatedReadmeMdFile,omitempty"`
 	RelatedReadmeMdFiles    []string                      `json:"relatedReadmeMdFiles,omitempty"`
 	InstallInstructionInput InstallInstructionScriptInput `json:"installInstructionInput,omitempty"`
 }
@@ -77,6 +78,7 @@ type PackageResult struct {
 	Version             string                          `json:"version,omitempty"`
 	PackageName         string                          `json:"packageName,omitempty"`
 	Path                []string                        `json:"path"`
+	PackageFolder       string                          `json:"packageFolder"`
 	ReadmeMd            []string                        `json:"readmeMd"`
 	Changelog           *Changelog                      `json:"changelog,omitempty"`
 	Artifacts           []string                        `json:"artifacts,omitempty"`