+++ sdk/azidentity/interactive_browser_credential.go
@@ -5,19 +5,12 @@ package azidentity
 
 import (
 	"context"
-	"crypto/sha256"
-	"encoding/base64"
 	"errors"
-	"math/rand"
-	"net/url"
-	"path"
-	"strings"
 
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/runtime"
-	"github.com/Azure/azure-sdk-for-go/sdk/internal/uuid"
-	"github.com/pkg/browser"
+	"github.com/AzureAD/microsoft-authentication-library-for-go/apps/public"
 )
 
 // InteractiveBrowserCredentialOptions contains optional parameters for InteractiveBrowserCredential.
@@ -49,8 +42,9 @@ func (o *InteractiveBrowserCredentialOptions) init() {
 
 // InteractiveBrowserCredential opens a browser to interactively authenticate a user.
 type InteractiveBrowserCredential struct {
-	client  *aadIdentityClient
+	client  publicClient
 	options InteractiveBrowserCredentialOptions
+	account public.Account
 }
 
 // NewInteractiveBrowserCredential constructs a new InteractiveBrowserCredential.
@@ -68,7 +62,10 @@ func NewInteractiveBrowserCredential(options *InteractiveBrowserCredentialOption
 	if err != nil {
 		return nil, err
 	}
-	c, err := newAADIdentityClient(authorityHost, &cp.ClientOptions)
+	c, err := public.New(cp.ClientID,
+		public.WithAuthority(runtime.JoinPaths(authorityHost, cp.TenantID)),
+		public.WithHTTPClient(newPipelineAdapter(&cp.ClientOptions)),
+	)
 	if err != nil {
 		return nil, err
 	}
@@ -79,75 +76,24 @@ func NewInteractiveBrowserCredential(options *InteractiveBrowserCredentialOption
 // ctx: Context used to control the request lifetime.
 // opts: Options for the token request, in particular the desired scope of the access token.
 func (c *InteractiveBrowserCredential) GetToken(ctx context.Context, opts policy.TokenRequestOptions) (*azcore.AccessToken, error) {
-	tk, err := c.client.authenticateInteractiveBrowser(ctx, &c.options, opts.Scopes)
+	ar, err := c.client.AcquireTokenSilent(ctx, opts.Scopes, public.WithSilentAccount(c.account))
+	if err == nil {
+		logGetTokenSuccess(c, opts)
+		return &azcore.AccessToken{Token: ar.AccessToken, ExpiresOn: ar.ExpiresOn.UTC()}, err
+	}
+
+	o := []public.InteractiveAuthOption{}
+	if c.options.RedirectURL != "" {
+		o = append(o, public.WithRedirectURI(c.options.RedirectURL))
+	}
+	ar, err = c.client.AcquireTokenInteractive(ctx, opts.Scopes, o...)
 	if err != nil {
 		addGetTokenFailureLogs("Interactive Browser Credential", err, true)
-		return nil, err
+		return nil, newAuthenticationFailedError(err, nil)
 	}
+	c.account = ar.Account
 	logGetTokenSuccess(c, opts)
-	return tk, nil
+	return &azcore.AccessToken{Token: ar.AccessToken, ExpiresOn: ar.ExpiresOn.UTC()}, err
 }
 
 var _ azcore.TokenCredential = (*InteractiveBrowserCredential)(nil)
-
-// authCodeReceiver is used to allow for testing without opening an interactive browser window. Allows mocking a response authorization code and redirect URI.
-var authCodeReceiver = func(ctx context.Context, authorityHost string, opts *InteractiveBrowserCredentialOptions, scopes []string) (*interactiveConfig, error) {
-	return interactiveBrowserLogin(ctx, authorityHost, opts, scopes)
-}
-
-// interactiveBrowserLogin opens an interactive browser with the specified tenant and client IDs provided then returns the authorization code
-// received or an error.
-func interactiveBrowserLogin(ctx context.Context, authorityHost string, opts *InteractiveBrowserCredentialOptions, scopes []string) (*interactiveConfig, error) {
-	// start local redirect server so login can call us back
-	rs := newServer()
-	state, err := uuid.New()
-	if err != nil {
-		return nil, err
-	}
-	redirectURL := rs.Start(state.String())
-	defer rs.Stop()
-	u, err := url.Parse(authorityHost)
-	if err != nil {
-		return nil, err
-	}
-	values := url.Values{}
-	values.Add("response_type", "code")
-	values.Add("response_mode", "query")
-	values.Add("client_id", opts.ClientID)
-	values.Add("redirect_uri", redirectURL)
-	values.Add("state", state.String())
-	values.Add("scope", strings.Join(scopes, " "))
-	values.Add("prompt", "select_account")
-	cv := ""
-	// the code verifier is a random 32-byte sequence that's been base-64 encoded without padding.
-	// it's used to prevent MitM attacks during auth code flow, see https://tools.ietf.org/html/rfc7636
-	b := make([]byte, 32) // nolint:gosimple
-	if _, err := rand.Read(b); err != nil {
-		return nil, err
-	}
-	cv = base64.RawURLEncoding.EncodeToString(b)
-	// for PKCE, create a hash of the code verifier
-	cvh := sha256.Sum256([]byte(cv))
-	values.Add("code_challenge", base64.RawURLEncoding.EncodeToString(cvh[:]))
-	values.Add("code_challenge_method", "S256")
-	u.Path = runtime.JoinPaths(u.Path, opts.TenantID, path.Join(oauthPath(opts.TenantID), "/authorize"))
-	u.RawQuery = values.Encode()
-	// open browser window so user can select credentials
-	if err = browser.OpenURL(u.String()); err != nil {
-		return nil, err
-	}
-	// now wait until the logic calls us back
-	if err := rs.WaitForCallback(ctx); err != nil {
-		return nil, err
-	}
-
-	authCode, err := rs.AuthorizationCode()
-	if err != nil {
-		return nil, err
-	}
-	return &interactiveConfig{
-		authCode:     authCode,
-		codeVerifier: cv,
-		redirectURI:  redirectURL,
-	}, nil
-}