+++ sdk/internal/recording/recording.go
@@ -10,6 +10,7 @@ import (
 	"bytes"
 	"crypto/tls"
 	"crypto/x509"
+	"encoding/json"
 	"errors"
 	"fmt"
 	"io/ioutil"
@@ -492,6 +493,7 @@ const (
 type recordedTest struct {
 	recordingId string
 	liveOnly    bool
+	variables   map[string]interface{}
 }
 
 var testSuite = map[string]recordedTest{}
@@ -505,6 +507,7 @@ var client = http.Client{
 type RecordingOptions struct {
 	UseHTTPS        bool
 	GroupForReplace string
+	Variables       map[string]interface{}
 }
 
 func defaultOptions() *RecordingOptions {
@@ -548,6 +551,7 @@ func getTestId(pathToRecordings string, t *testing.T) string {
 	return path.Join(pathToRecordings, "recordings", t.Name()+".json")
 }
 
+// Start tells the test proxy to begin accepting requests for a given test
 func Start(t *testing.T, pathToRecordings string, options *RecordingOptions) error {
 	if options == nil {
 		options = defaultOptions()
@@ -580,20 +584,42 @@ func Start(t *testing.T, pathToRecordings string, options *RecordingOptions) err
 	recId := resp.Header.Get(IDHeader)
 	if recId == "" {
 		b, err := ioutil.ReadAll(resp.Body)
+		defer resp.Body.Close()
 		if err != nil {
 			return err
 		}
 		return fmt.Errorf("Recording ID was not returned by the response. Response body: %s", b)
 	}
+
+	// Unmarshal any variables returned by the proxy
+	var m map[string]interface{}
+	body, err := ioutil.ReadAll(resp.Body)
+	defer resp.Body.Close()
+	if err != nil {
+		return err
+	}
+	if len(body) > 0 {
+		err = json.Unmarshal(body, &m)
+		if err != nil {
+			return err
+		}
+	}
+
 	if val, ok := testSuite[t.Name()]; ok {
 		val.recordingId = recId
+		val.variables = m
 		testSuite[t.Name()] = val
 	} else {
-		testSuite[t.Name()] = recordedTest{recordingId: recId, liveOnly: false}
+		testSuite[t.Name()] = recordedTest{
+			recordingId: recId,
+			liveOnly:    false,
+			variables:   m,
+		}
 	}
 	return nil
 }
 
+// Stop tells the test proxy to stop accepting requests for a given test
 func Stop(t *testing.T, options *RecordingOptions) error {
 	if options == nil {
 		options = defaultOptions()
@@ -614,13 +640,32 @@ func Stop(t *testing.T, options *RecordingOptions) error {
 	if err != nil {
 		return err
 	}
+	if len(options.Variables) > 0 {
+		req.Header.Set("Content-Type", "application/json")
+		marshalled, err := json.Marshal(options.Variables)
+		if err != nil {
+			return err
+		}
+		req.Body = ioutil.NopCloser(bytes.NewReader(marshalled))
+		req.ContentLength = int64(len(marshalled))
+	}
+
 	var recTest recordedTest
 	var ok bool
 	if recTest, ok = testSuite[t.Name()]; !ok {
 		return errors.New("Recording ID was never set. Did you call StartRecording?")
 	}
 	req.Header.Set("x-recording-id", recTest.recordingId)
-	_, err = client.Do(req)
+	resp, err := client.Do(req)
+	if resp.StatusCode != 200 {
+		b, err := ioutil.ReadAll(resp.Body)
+		defer resp.Body.Close()
+		if err == nil {
+			return fmt.Errorf("proxy did not stop the recording properly: %s", string(b))
+		}
+		return fmt.Errorf("proxy did not stop the recording properly: %s", err.Error())
+	}
+	_ = resp.Body.Close()
 	return err
 }
 
@@ -687,6 +732,7 @@ func (c RecordingHTTPClient) Do(req *http.Request) (*http.Response, error) {
 	return c.defaultClient.Do(req)
 }
 
+// NewRecordingHTTPClient returns a type that implements `azcore.Transporter`. This will automatically route tests on the `Do` call.
 func NewRecordingHTTPClient(t *testing.T, options *RecordingOptions) (*RecordingHTTPClient, error) {
 	if options == nil {
 		options = &RecordingOptions{UseHTTPS: true}
@@ -720,4 +766,12 @@ func IsLiveOnly(t *testing.T) bool {
 		return s.liveOnly
 	}
 	return false
+}
+
+// GetVariables returns access to the variables stored by the test proxy for a specific test
+func GetVariables(t *testing.T) map[string]interface{} {
+	if s, ok := testSuite[t.Name()]; ok {
+		return s.variables
+	}
+	return nil
 }